requirements.txt
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
python-dotenv==1.0.1